Reorganize code with components and routes
